append_claim, append_evidenceの実装と移動
