+++ src/services/evidence.ts
@@ -3,9 +3,11 @@ import {evidence_slice} from 'stores/slices/evidence';
 import {generate_evidence_id} from 'stores/ids/id_generators';
 import {baseModel} from 'models/baseModel';
 import {mEvidence, mEvidenceSignature} from 'models/mEvidence';
-import { get_from_id } from './id';
+import { get_from_id, get_parent_id } from './id';
 import { is_mPoint } from 'models/mPoint';
 import { point_slice } from 'stores/slices/point';
+import { part_add_child } from './part';
+import { switch_for_append, point_add_child } from './point';
 
 export const generate_evidence=(
   parent: baseModel['id'],
@@ -22,4 +24,16 @@ export const generate_evidence=(
   store.dispatch(evidence_slice.actions.add(generated));
   store.dispatch(point_slice.actions.addChild([parent, generated.id]));
   return generated;
-}
+};
+
+export const append_evidence=(parent_id: baseModel['id']): mEvidence|undefined=>{
+  return switch_for_append(
+    parent_id,
+    (id)=>append_evidence(part_add_child(id).id),
+    (id)=>{
+      const parent_id=get_parent_id(id);
+      if(parent_id!==undefined&&parent_id!==null) return append_evidence(parent_id);
+    },
+    (id)=>point_add_child(id,mEvidenceSignature)
+  );
+};
